create getCatalog function. remove catalog from UI when new data loading
